requirements.txt
jupyter==1.0.0
pytubefix==6.9.2
moviepy==1.0.3
langchain==0.2.12
langchain-text-splitters==0.2.2
langchain-core==0.2.29
langchain-community==0.2.11
lancedb==0.12.0
langchain-openai==0.1.21
tiktoken==0.7.0
openai==1.40.3
python-dotenv==1.0.1
azure-cognitiveservices-speech==1.38.0